Перенос тестов nonCommand в один класс
